requirements.txt
gradio
openai
wikipedia
twitter-api-client
python-dotenv
pytest
youtube-transcript-api
py-trees
python-pptx
python-docx
markdown
moviepy
opencv-python
scipy
prefect
googlemaps
